+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [App4.1.0.beta6] [ShortcutsPlugin1.0.0] - 2025-11-24
+### Added
+- Shortcuts plugin: create custom buttons that load a chosen N.I.N.A. sequence and optionally auto-start it with one tap.
+- Add feature highlight
+
 ## [App4.1.0.beta5] - 2025-11-24
 ### Added
 - PHD2 looping